+++ packages/core/src/__tests__/lint.test.ts
@@ -6,7 +6,6 @@ import { describe, it, expect } from 'vitest';
 
 import { parseYamlToDocument, replaceSourceWithRef } from '../../__tests__/utils.js';
 import { createConfig, loadConfig, loadIgnoreConfig } from '../config/load.js';
-import { detectSpec } from '../detect-spec.js';
 import { lintFromString, lintConfig, lintDocument, lint } from '../lint.js';
 import { BaseResolver } from '../resolve.js';
 import { createConfigTypes } from '../types/redocly-yaml.js';
@@ -1636,48 +1635,6 @@ describe('lint', () => {
     });
 
     expect(replaceSourceWithRef(results)).toMatchInlineSnapshot(`[]`);
-  });
-
-  it('detect OpenAPI should throw an error when version is not string', () => {
-    const testDocument = parseYamlToDocument(
-      outdent`
-        openapi: 3.0
-      `,
-      ''
-    );
-    expect(() => detectSpec(testDocument.parsed)).toThrow(
-      `Invalid OpenAPI version: should be a string but got "number"`
-    );
-  });
-
-  it('detect unsupported OpenAPI version', () => {
-    const testDocument = parseYamlToDocument(
-      outdent`
-        openapi: 1.0.4
-      `,
-      ''
-    );
-    expect(() => detectSpec(testDocument.parsed)).toThrow(`Unsupported OpenAPI version: 1.0.4`);
-  });
-
-  it('detect unsupported AsyncAPI version', () => {
-    const testDocument = parseYamlToDocument(
-      outdent`
-        asyncapi: 1.0.4
-      `,
-      ''
-    );
-    expect(() => detectSpec(testDocument.parsed)).toThrow(`Unsupported AsyncAPI version: 1.0.4`);
-  });
-
-  it('detect unsupported spec format', () => {
-    const testDocument = parseYamlToDocument(
-      outdent`
-        notapi: 3.1.0
-      `,
-      ''
-    );
-    expect(() => detectSpec(testDocument.parsed)).toThrow(`Unsupported specification`);
   });
 
   it("struct rule shouldn't throw an error for named callback", async () => {
